refactor YjsService initialization
